==> analises/15_cleaning-scraped-data.ipynb <==
# Cleaning and exploring data

import pandas as pd
from glob import glob

import xavy.explore as xe
import xavy.dataframes as xd

## Load data

filenames = glob('../dados/brutos/w3c-wg-members_*.csv')
raw_members_df = pd.concat([pd.read_csv(f) for f in filenames])

## Explore raw data

**Findings**
1. Organization name and URL are biunivocal
2. The only organization with no URL is W3C or its invited experts
3. There are people affiliated to different organizations
4. Invited experts may come from companies
5. E-mails identify people, but people can have more than one email:
    1. All members have a listed email.
    2. Ignoring people with no picture, all pictures are biunivocal to emails
    3. There are no emails with more than one name associated to it.
    4. **There are people with more than one email**

### Basic stuff

xe.checkMissing(raw_members_df)

xe.mapUnique(raw_members_df)

### Sanity checks

#### Organization URL

xe.one2oneViolations(raw_members_df, 'org_name', 'org_url')

xe.one2oneViolations(raw_members_df, 'org_url', 'org_name')

raw_members_df.loc[raw_members_df['org_url'].isnull(), 'org_name'].value_counts()

#### Picture

picture_df = raw_members_df.loc[raw_members_df['picture'] != 'https://www.w3.org/assets/website-2021/svg/avatar.svg']

xe.one2oneQ(picture_df, 'email', 'picture')

#### Email

xd.print_array_series(xe.one2oneViolations(raw_members_df, 'name', 'email'))

xe.one2oneViolations(raw_members_df, 'email', 'name')

#### Affiliation

xd.print_array_series(xe.one2oneViolations(raw_members_df, 'name', 'org_name'))

raw_members_df

raw_members_df.query('role != "staff"')['org_name'].value_counts(normalize=True).head(20)

raw_members_df.query('role == "chairs"')['org_name'].value_counts(normalize=True).head(20)

# Trash

groups_df = pd.read_csv('../dados/brutos/w3c_wg_2024-07-04.csv')

groups_df
