==> analises/20_getting-data-from-api.ipynb <==
# Getting data from W3C API

import requests
import json
import pandas as pd
import matplotlib.pyplot as pl
from datetime import date
from glob import glob
from bs4 import BeautifulSoup

import xavy.explore as xe
import xavy.dataframes as xd

import modules.w3c_api as wa

## Capturing data

Queremos a lista de instâncias -- com seus detalhes -- de cada uma das categorias:

1. groups
2. participants 
3. affiliations (organizations)

Também queremos as conexões:

1. groups - participants
2. participants - affiliations

### Groups

#### List

# Get list of all groups:
#group_types = {'bg': 'Business Group', 'other':'Permanent Group', 'ig':'Interest Group', 'wg': 'Working Group', 'tf': 'Task Force', 'cg': 'Community Group'}
groups = wa.get_all_data('https://api.w3.org/groups')
groups_df = pd.DataFrame(groups)
#groups_df['type_id'] = groups_df['href'].str.split('/').str.slice(4,5).str.join('')
#groups_df['type'] = groups_df['type_id'].map(group_types)
# Make sure group title is unique:
#assert groups_df['title'].nunique() == len(groups_df)

groups_df

# Save data:
today = date.today().strftime('%Y-%m-%d')
#groups_df.to_csv('../data/raw/api/w3c_groups_list_{}.csv'.format(today), index=False)

#### Properties

# Get properties from all groups:
group_extras = ['homepage', 'users', 'participations', 'chairs']
groups_properties = [wa.get_properties(groups_df.iloc[i]['href'], group_extras) for i in range(len(groups_df))]
gproperties_df = pd.DataFrame(groups_properties)

gproperties_df

# Save data:
today = date.today().strftime('%Y-%m-%d')
#gproperties_df.to_csv('../data/raw/api/w3c_groups_properties_{}.csv'.format(today), index=False)

#### Links to Participants

# Select groups with users:
groups_w_users_df = gproperties_df.dropna(subset='users')

# Loop over groups:
for i in range(0, len(groups_w_users_df)):
    print(i, end=' ')
    # Select group:
    group_card = groups_w_users_df.iloc[i]
    # Get list of users:
    endpoint = group_card['users']
    group_members = wa.get_all_data(endpoint)
    group_members_df = pd.DataFrame(group_members)
    group_members_df['call'] = endpoint
    # Save:
    today = date.today().strftime('%Y-%m-%d')
    gid = wa.unique_shortname(group_card['users'])
    #group_members_df.to_csv('../data/raw/api/w3c_participants_group-{:}_{:}.csv'.format(gid, today), index=False)

#### Links to chairs

# Select groups with users:
groups_w_chairs_df = gproperties_df.dropna(subset='chairs')

# Loop over groups:
for i in range(0, len(groups_w_chairs_df)):
    print(i, end=' ')
    # Select group:
    group_card = groups_w_chairs_df.iloc[i]
    # Get list of users:
    endpoint = group_card['chairs']
    group_chairs = wa.get_all_data(endpoint)
    group_chairs_df = pd.DataFrame(group_chairs)
    group_chairs_df['call'] = endpoint
    # Save:
    today = date.today().strftime('%Y-%m-%d')
    gid = wa.unique_shortname(group_card['chairs'])
    #group_chairs_df.to_csv('../data/raw/api/w3c_chairs_group-{:}_{:}.csv'.format(gid, today), index=False)

### Participants' affiliations

# Load raw data about group participants:
group_participants_df = wa.load_raw_api_data('../data/raw/api/w3c_participants_group-*_2024-07-25.csv')

# Users affiliations endpoints:
aff_endpoints = group_participants_df['href'].drop_duplicates() + '/affiliations'

# Capture parameters:
batch_size = 100
# Derived parameters:
n_users = len(aff_endpoints)
n_batches = int(n_users / batch_size) + 1
n_digits = len(str(n_batches))

# Loop over batches:
for b in range(0, n_batches):
    # Batch label:
    batch_label = str(b).zfill(n_digits)
    print(batch_label, end=' ')

    # Batch limits:
    b_start = b * batch_size
    b_end   = (b + 1) * batch_size

    # Loop over users (group participants):
    affs_dfs = [ wa.get_user_affiliations(aff_endpoints.iloc[i]) for i in range(b_start, min(b_end, n_users)) ]

    # Save batch data:
    today = date.today().strftime('%Y-%m-%d')
    users_affs_df = pd.concat(affs_dfs, ignore_index=True)
    #users_affs_df.to_csv('../data/raw/api/w3c_affiliations_batch-{:}_{:}.csv'.format(batch_label, today), index=False)

# Link para organização e seu nome são identificadores:
assert xe.one2oneQ(users_affs_df, 'href', 'title')
# Cada conexão entre usuário e organização só aparece uma vez:
assert xd.iskeyQ(users_affs_df)

### Organizations' details

# Load raw data about participants' affiliations:
affiliations_df = wa.load_raw_api_data('../data/raw/api/w3c_affiliations_batch-*_2024-07-25.csv')

# Endpoints with the organizations' details:
org_endpoints = affiliations_df['href'].drop_duplicates()

# Capture parameters:
batch_size = 100
# Derived parameters:
n_orgs = len(org_endpoints)
n_batches = int(n_orgs / batch_size) + 1
n_digits = len(str(n_batches))

# Loop over batches:
for b in range(0, n_batches):
    # Batch label:
    batch_label = str(b).zfill(n_digits)
    print(batch_label, end=' ')

    # Batch limits:
    b_start = b * batch_size
    b_end   = (b + 1) * batch_size

    # Loop over organizations (participants affiliations):
    orgs = [wa.get_properties(org_endpoints.iloc[i], []) for i in range(b_start, min(b_end, n_orgs)) ]

    # Save batch data:
    today = date.today().strftime('%Y-%m-%d')
    orgs_df = pd.DataFrame(orgs)
    #orgs_df.to_csv('../data/raw/api/w3c_organizations_batch-{:}_{:}.csv'.format(batch_label, today), index=False)

### Affiliations

#### List

affiliations = wa.get_all_data('https://api.w3.org/affiliations')
affiliations_df = pd.DataFrame(affiliations)

affiliations_df

# Save data:
today = date.today().strftime('%Y-%m-%d')
#affiliations_df.to_csv('../data/raw/api/w3c_organizations_list_{}.csv'.format(today), index=False)

#### Properties

# Capture parameters:
org_extras = ['homepage', 'participants', 'participations', 'logo']
batch_size = 100
# Derived parameters:
n_orgs = len(affiliations_df)
n_batches = int(n_orgs / batch_size) + 1
n_digits = len(str(n_batches))

# Loop over batches:
for b in range(0, n_batches):
    # Batch label:
    batch_label = str(b).zfill(n_digits)
    print(batch_label, end=' ')

    # Batch limits:
    b_start = b * batch_size
    b_end   = (b + 1) * batch_size

    # Loop over organizations:
    org_properties = [wa.get_properties(affiliations_df.iloc[i]['href'], org_extras, True) for i in range(b_start, min(b_end, n_orgs))]

    # Save batch data:
    today = date.today().strftime('%Y-%m-%d')
    oproperties_df = pd.DataFrame(org_properties)
    oproperties_df.to_csv('../data/raw/api/w3c_organizations_batch-{:}_{:}.csv'.format(batch_label, today), index=False)

### Participants

#### List

# Load organization files:
org_files = glob('../data/raw/api/w3c_organizations_batch-*_2024-07-06.csv')
org_details_df = pd.concat([pd.read_csv(f) for f in org_files], ignore_index=True)
# Remove 404 NOT FOUND organizations:
existing_orgs_df = org_details_df.loc[~org_details_df['participants'].isnull()]

for i in range(len(existing_orgs_df)):
    # Get list of members in the organization:
    endpoint = existing_orgs_df.iloc[i]['participants']
    members = wa.get_all_data(endpoint)
    if len(members) > 0:
        print(i, end=' ')
        # Put in DataFrame:
        members_df = pd.DataFrame(members)
        members_df['call'] = endpoint
        # Save organization members data:
        org_id = endpoint.split('/')[-2]
        today = date.today().strftime('%Y-%m-%d')
        members_df.to_csv('../data/raw/api/w3c_participants_org-{:}_{:}.csv'.format(org_id, today), index=False)

# Load members files:
members_files = glob('../data/raw/api/w3c_participants_org-*_2024-07-06.csv')
members_df = pd.concat([pd.read_csv(f) for f in members_files], ignore_index=True)

## Documentation

# Load documentation from API:
response = requests.get('https://api.w3.org/doc')
soup = BeautifulSoup(response.text)
scripts = soup.find_all('script')
doc_json = json.loads(scripts[0].text)

spec = doc_json['spec']
endpoints = spec['paths']
endpoints['/groups/{group}']

spec['definitions'].keys()
